+++ packages/react-native-the-sheet/src/core/backdrop/types.ts
@@ -1,6 +1,9 @@
 import type { StyleProp, ViewStyle } from 'react-native'
+import type { AnimatedProp } from '../types'
 
 export type BackdropProps = {
+  disabled?: AnimatedProp<boolean>
+
   styles?: {
     root?: StyleProp<ViewStyle>
   }