+++ packages/react-native-the-sheet/src/core/bottom-sheet/hooks/index.ts
@@ -1,2 +1,4 @@
+export * from './use-bottom-sheet-pan-gesture-with-lock-scroll'
 export * from './use-bottom-sheet-pan-gesture'
 export * from './use-bottom-sheet-scroll-view-utils'
+export * from './use-pan-gesture-lock-scroll'