+++ packages/react-native-the-sheet/src/core/bottom-sheet/hooks/use-pan-gesture-lock-scroll.tsx
@@ -1,7 +1,7 @@
 import { scrollTo, useAnimatedReaction } from 'react-native-reanimated'
-import { isApproxEqual } from '../../../private/utils/approximately-equal'
+import { isApproxEqual } from '../../utils/approximately-equal'
 import { TRANSLATE_Y_REST_THRESHOLD } from '../constants'
-import { useBottomSheet } from '../../bottom-sheet-provider'
+import { useBottomSheet } from '../bottom-sheet-provider'
 
 export const usePanGestureLockScroll = () => {
   const {